mypy

+++ ffcx/codegeneration/common.py
@@ -60,7 +60,7 @@ class KernelTensorSizes(NamedTuple):
 
 def integral_tensor_sizes(ir: IntegralIR) -> KernelTensorSizes:
     """Compute tensor sizes of integral IR input data."""
-    A = np.prod(ir.expression.tensor_shape)
+    A = np.prod(ir.expression.tensor_shape, dtype=int)
     w = sum(coeff.ufl_element().dim for coeff in ir.expression.coefficient_offsets.keys())
     c = sum(
         np.prod(constant.ufl_shape, dtype=int)
@@ -78,7 +78,7 @@ def expression_tensor_sizes(ir: ExpressionIR) -> KernelTensorSizes:
     num_points = next(iter(ir.expression.integrand))[1].points.shape[0]
     num_components = np.prod(ir.expression.shape, dtype=np.int32)
     num_argument_dofs = np.prod(ir.expression.tensor_shape, dtype=np.int32)
-    A = int(num_points * num_components * num_argument_dofs)  # ref. ufcx.h
+    A = num_points * num_components * num_argument_dofs  # ref. ufcx.h
     w = sum(coeff.ufl_element().dim for coeff in ir.expression.coefficient_offsets.keys())
     c = sum(
         np.prod(constant.ufl_shape, dtype=int)
